replace get_failover_ip by is_failover_ip_routed_to_me
Hetzner API is limited to 100 requests per hour what limits the monitoring of Hetzner
failover IP address. So I implemented the monitoring through fping which is capable
to figure out if the failover IP is routed to us or not.

Some older versions of fping does not support -S option so be sure that your version
up2date enough.

Signed-off-by: Pavel Pulec <kayn@inuits.eu>